feat: add Cypress tests
